Refs #32018. Review changes.
