+++ docs/source/release/v6.2.0/muon.rst
@@ -61,5 +61,6 @@ Algorithms
 ##########
 
 - Updated :ref:`LoadMuonLog <algm-LoadMuonLog>` to read units for most log values.
+- It is now possible to exclude a fit range when executing the :ref:`CalculateMuonAsymmetry <algm-CalculateMuonAsymmetry>` algorithm.
 
 :ref:`Release 6.2.0 <v6.2.0>`
