+++ scripts/test/Muon/fitting_widgets/basic_fitting/basic_fitting_presenter_test.py
@@ -9,17 +9,10 @@ from unittest import mock
 
 from mantid.api import FrameworkManager, FunctionFactory
 from mantidqt.widgets.fitscriptgenerator import FittingMode
-
-from Muon.GUI.Common.fitting_widgets.basic_fitting.basic_fitting_model import BasicFittingModel
-from Muon.GUI.Common.fitting_widgets.basic_fitting.basic_fitting_presenter import BasicFittingPresenter
-from Muon.GUI.Common.fitting_widgets.basic_fitting.basic_fitting_view import BasicFittingView
-from Muon.GUI.Common.test_helpers.fitting_mock_setup import (add_mock_methods_to_basic_fitting_model,
-                                                             add_mock_methods_to_basic_fitting_presenter,
-                                                             add_mock_methods_to_basic_fitting_view)
-from Muon.GUI.Common.thread_model import ThreadModel
+from Muon.GUI.Common.test_helpers.fitting_mock_setup import MockBasicFitting
 
 
-class BasicFittingPresenterTest(unittest.TestCase):
+class BasicFittingPresenterTest(unittest.TestCase, MockBasicFitting):
 
     @classmethod
     def setUpClass(cls):
@@ -444,117 +437,6 @@ class BasicFittingPresenterTest(unittest.TestCase):
         self.mock_view_start_x.assert_called_once_with(self.start_x)
         self.mock_model_current_end_x.assert_called_once_with()
         self.mock_view_end_x.assert_called_once_with(self.end_x)
-
-    def _setup_mock_view(self):
-        self.view = mock.Mock(spec=BasicFittingView)
-        self.view = add_mock_methods_to_basic_fitting_view(self.view)
-
-        # Mock the properties of the view
-        self.mock_view_minimizer = mock.PropertyMock(return_value=self.minimizer)
-        type(self.view).minimizer = self.mock_view_minimizer
-        self.mock_view_evaluation_type = mock.PropertyMock(return_value=self.evaluation_type)
-        type(self.view).evaluation_type = self.mock_view_evaluation_type
-        self.mock_view_fit_to_raw = mock.PropertyMock(return_value=self.fit_to_raw)
-        type(self.view).fit_to_raw = self.mock_view_fit_to_raw
-        self.mock_view_fit_object = mock.PropertyMock(return_value=self.fit_function)
-        type(self.view).fit_object = self.mock_view_fit_object
-        self.mock_view_start_x = mock.PropertyMock(return_value=self.start_x)
-        type(self.view).start_x = self.mock_view_start_x
-        self.mock_view_end_x = mock.PropertyMock(return_value=self.end_x)
-        type(self.view).end_x = self.mock_view_end_x
-        self.mock_view_exclude_range = mock.PropertyMock(return_value=True)
-        type(self.view).exclude_range = self.mock_view_exclude_range
-        self.mock_view_exclude_start_x = mock.PropertyMock(return_value=self.start_x)
-        type(self.view).exclude_start_x = self.mock_view_exclude_start_x
-        self.mock_view_exclude_end_x = mock.PropertyMock(return_value=self.end_x)
-        type(self.view).exclude_end_x = self.mock_view_exclude_end_x
-        self.mock_view_plot_guess = mock.PropertyMock(return_value=self.plot_guess)
-        type(self.view).plot_guess = self.mock_view_plot_guess
-        self.mock_view_function_name = mock.PropertyMock(return_value=self.function_name)
-        type(self.view).function_name = self.mock_view_function_name
-
-    def _setup_mock_model(self):
-        self.model = mock.Mock(spec=BasicFittingModel)
-        self.model = add_mock_methods_to_basic_fitting_model(self.model, self.dataset_names, self.current_dataset_index,
-                                                             self.fit_function, self.start_x, self.end_x,
-                                                             self.fit_status, self.chi_squared)
-
-        # Mock the properties of the model
-        self.mock_model_current_dataset_index = mock.PropertyMock(return_value=self.current_dataset_index)
-        type(self.model).current_dataset_index = self.mock_model_current_dataset_index
-        self.mock_model_dataset_names = mock.PropertyMock(return_value=self.dataset_names)
-        type(self.model).dataset_names = self.mock_model_dataset_names
-        self.mock_model_current_dataset_name = mock.PropertyMock(return_value=
-                                                                 self.dataset_names[self.current_dataset_index])
-        type(self.model).current_dataset_name = self.mock_model_current_dataset_name
-        self.mock_model_number_of_datasets = mock.PropertyMock(return_value=len(self.dataset_names))
-        type(self.model).number_of_datasets = self.mock_model_number_of_datasets
-        self.mock_model_start_xs = mock.PropertyMock(return_value=[self.start_x] * len(self.dataset_names))
-        type(self.model).start_xs = self.mock_model_start_xs
-        self.mock_model_current_start_x = mock.PropertyMock(return_value=self.start_x)
-        type(self.model).current_start_x = self.mock_model_current_start_x
-        self.mock_model_end_xs = mock.PropertyMock(return_value=[self.end_x] * len(self.dataset_names))
-        type(self.model).end_xs = self.mock_model_end_xs
-        self.mock_model_current_end_x = mock.PropertyMock(return_value=self.end_x)
-        type(self.model).current_end_x = self.mock_model_current_end_x
-        self.mock_model_exclude_range = mock.PropertyMock(return_value=True)
-        type(self.model).exclude_range = self.mock_model_exclude_range
-        self.mock_model_current_exclude_start_x = mock.PropertyMock(return_value=self.start_x)
-        type(self.model).current_exclude_start_x = self.mock_model_current_exclude_start_x
-        self.mock_model_current_exclude_end_x = mock.PropertyMock(return_value=self.end_x)
-        type(self.model).current_exclude_end_x = self.mock_model_current_exclude_end_x
-        self.mock_model_plot_guess = mock.PropertyMock(return_value=self.plot_guess)
-        type(self.model).plot_guess = self.mock_model_plot_guess
-        self.mock_model_minimizer = mock.PropertyMock(return_value=self.minimizer)
-        type(self.model).minimizer = self.mock_model_minimizer
-        self.mock_model_evaluation_type = mock.PropertyMock(return_value=self.evaluation_type)
-        type(self.model).evaluation_type = self.mock_model_evaluation_type
-        self.mock_model_fit_to_raw = mock.PropertyMock(return_value=self.fit_to_raw)
-        type(self.model).fit_to_raw = self.mock_model_fit_to_raw
-        self.mock_model_single_fit_functions = mock.PropertyMock(return_value=self.single_fit_functions)
-        type(self.model).single_fit_functions = self.mock_model_single_fit_functions
-        self.mock_model_current_single_fit_function = mock.PropertyMock(return_value=self.fit_function)
-        type(self.model).current_single_fit_function = self.mock_model_current_single_fit_function
-        self.mock_model_single_fit_functions_cache = mock.PropertyMock(return_value=self.fit_function)
-        type(self.model).single_fit_functions_cache = self.mock_model_single_fit_functions_cache
-        self.mock_model_fit_statuses = mock.PropertyMock(return_value=[self.fit_status] * len(self.dataset_names))
-        type(self.model).fit_statuses = self.mock_model_fit_statuses
-        self.mock_model_current_fit_status = mock.PropertyMock(return_value=self.fit_status)
-        type(self.model).current_fit_status = self.mock_model_current_fit_status
-        self.mock_model_chi_squared = mock.PropertyMock(return_value=[self.chi_squared] * len(self.dataset_names))
-        type(self.model).chi_squared = self.mock_model_chi_squared
-        self.mock_model_current_chi_squared = mock.PropertyMock(return_value=self.chi_squared)
-        type(self.model).current_chi_squared = self.mock_model_current_chi_squared
-        self.mock_model_function_name = mock.PropertyMock(return_value=self.function_name)
-        type(self.model).function_name = self.mock_model_function_name
-        self.mock_model_function_name_auto_update = mock.PropertyMock(return_value=True)
-        type(self.model).function_name_auto_update = self.mock_model_function_name_auto_update
-        self.mock_model_simultaneous_fitting_mode = mock.PropertyMock(return_value=False)
-        type(self.model).simultaneous_fitting_mode = self.mock_model_simultaneous_fitting_mode
-        self.mock_model_global_parameters = mock.PropertyMock(return_value=[])
-        type(self.model).global_parameters = self.mock_model_global_parameters
-        self.mock_model_do_rebin = mock.PropertyMock(return_value=False)
-        type(self.model).do_rebin = self.mock_model_do_rebin
-
-    def _setup_presenter(self):
-        self.presenter = BasicFittingPresenter(self.view, self.model)
-        self.presenter = add_mock_methods_to_basic_fitting_presenter(self.presenter)
-
-        # Mock the fit result property
-        self.presenter.fitting_calculation_model = mock.Mock(spec=ThreadModel)
-        self.mock_presenter_get_fit_results = mock.PropertyMock(return_value=(self.fit_function, self.fit_status,
-                                                                              self.chi_squared))
-        type(self.presenter.fitting_calculation_model).result = self.mock_presenter_get_fit_results
-
-        # Mock unimplemented methods and notifiers
-        self.presenter.handle_fitting_finished = mock.Mock()
-        self.presenter.update_and_reset_all_data = mock.Mock()
-        self.presenter.disable_editing_notifier.notify_subscribers = mock.Mock()
-        self.presenter.enable_editing_notifier.notify_subscribers = mock.Mock()
-        self.presenter.selected_fit_results_changed.notify_subscribers = mock.Mock()
-        self.presenter.fit_function_changed_notifier.notify_subscribers = mock.Mock()
-        self.presenter.fit_parameter_changed_notifier.notify_subscribers = mock.Mock()
-        self.presenter.update_exclude_start_and_end_x_in_view_and_model = mock.Mock()
 
 
 if __name__ == '__main__':
